command line parser cleanup
Changed the --version option on the post command to --pversion so that
we can use --version for the program's version number.
Removed  the short option, -P, for --product from the search command
since that was causing a conflict.
Added a help command which will display the sub commands.  The --help
option is used with commands to display detailed help on each command.
Removed laxoptionparser since it is no longer needed.