Add --batch option to post command
The --batch option replaces the --no-append-command and --no-confirm
options.  If the --batch option is present on the command line, pybugz
expects all values needed to fill in the bug to be on the command line
and does not prompt for any options.  If a mandatory option is missing,
it gives an error message and aborts.  If an optional option is missing,
it will be set to its default value.