Fix auth search for (newer) bugzillas
Newer bugzillas tend to have modified title lines (e.g.
bugzilla.kernel.org: Log in to Kernel Bug Tracker) which would not be
picked up by Bugz.auth(), returning the "Already logged in" message.

Search for the "Log in to" string instead, as it's more likely to appear
in all bugzilla installations.