use argparse to handle command line parsing
This breaks backward compatibility, but it makes the command line
parsing code much cleaner and more flexable.

The difference for the user is that the global options must appear
before the sub command on the command line.

bugz -h

will give the list of global options and

bugz [subcommand] -h

will give the list of options which are specific to the subcommand.