Issue #12: Add support for defaults in the configuration file
There's a new (optional) section named [default], which can contain any of the
keys that other sections can. These are used by default, and overridden by the
explicitly named sections.

A special key named 'connection' may also be present, which sets the name of
the default section to use when no --connection parameter is used on the
command line.