Add workaround for bugzilla bug 4766 - 'Suspicious Action'
Bugzilla 3.2.2 has a bug https://bugzilla.mozilla.org/show_bug.cgi?id=4766,
where the authentication 'token' is sent in HTML output, but not XML output.
This means that bug 'modify' operations would appear to succeed, but actually
be failing with a 'Suspicious Action Detected' message. Quite misleading, and
it makes pybugz mostly useless with bugzilla 3.2.2.

This patch detects this issue, logs a warning, and works around the bug by
scraping the HTML for the token, and re-submitting the request.