Use xmlrpc interface
Replace the old Bugzilla interface with code that sets up a proxy so we
can communicate with Bugzilla through its xmlrpc interface.

This is much more stable than the previous interface to Bugzilla, and it
is the documented way for client software to communicate with Bugzilla.

I would like to thank Mike Gilbert <floppym@gentoo.org> for contributing
the Bugzilla Proxy code.