remove hard coded config.py
This file is no longer needed. If we need to find out the fields and
possible values, there are ways to query this information through the
xmlrpc interface.

Unfortunately this means that the --show-url option for the namedcmd and
search commands no longer works since we have no way of knowing what the
URL should be.