Implement new configuration system and connection object
Add a configuration system to pybugz which reads configuration files
as follows:

- pybugz-provided configuration files are stored in
  sys.prefix+'/share/pybugz.d/*.conf'.
- The system administrator can add configuration files in
  /etc/pybugz.d/*.conf.
- An individual user can add a configuration file. This file is named
  ~/.bugzrc by default, but this can be overridden by the --config-file
  command line option.

In each directory, files are read in lexical order. Later files can add
to or override definitions in earlier files. This means the system
administrator can modify the pybugz-supplied defaults, and a user can,
in turn, modify all settings.

The new connection object is used to hold all data parsed from the
command line and configuration files. The settings on the command line
take priority over all configuration settings.

This commit also begins adding support for debug logging by logging the
values stored in the connection object once it is set up.

I would like to thank Pavel Raiskup <praiskup@redhat.com> for the idea
for this system.