Migrate old token file into the new system
This migrates the old single token file to the new token configuration
file.

The old token will be assigned to the default connection, or the one you
use with the --connection switch when the old file is detected.

If this is not correct, you will need to edit ~/.bugz_tokens and change
the section name to match the appropriate connection.

This is also related to #78.