--not-status flag
I commonly want to see anything apart from unconfirmed or resolved bugs.
This is useful this.

I have also (somewhat contentiously) allowed a search without any
keyword specified. This is useful for getting all bugz for
a particularly product or component.

I don't know if we want to insist that *some* constraint is
applied: If people want to list all bugz they should
be probably be free to do so...