Add --interactive command line switch for username and password prompts
This switch allows you to decide whether or not  to be prompted for a
username and password if no credentials are configured for the bugzilla
you are using.

The default setting is not to prompt, which will mean that pybugz will
access bugzilla in read-only mode if no credentials are configured.