utils: use builtin way to get terminal width
`shutil.get_terminal_size` is available in Python 3.3 and later, and is
exactly what we need here. I'm unsure from where `win32utils` come from,
but there is no import for it or something, so it can't work.

I've verified `shutil.get_terminal_size` implementation on unix and it
does the same we did before, only better - so I see only gains from this
replace.

Resolves: https://github.com/williamh/pybugz/issues/114